core/services/pipeline: consolidate required paths check on BridgeTask
